Configure cron to synchronize channels every 30 minutes
At this moment I didn't want to add complexity on the system so I decided to use
a simple cron task to synchronize the channels. It was a little trickier as cron
don't handle the environment variables defined in docker-compose.yml very well.

In future, probably we will need a more robust solution like Celery tasks or
something similar to handle this task.